Convert handler tests to use Testify suites.
