+++ Gopkg.toml
@@ -3,7 +3,8 @@ required = [
 	"github.com/golang/lint/golint",
 	"github.com/go-swagger/go-swagger/cmd/swagger",
 	"github.com/segmentio/chamber",
-	"github.com/codegangsta/gin"
+	"github.com/codegangsta/gin",
+	"github.com/stretchr/testify/suite"
 	]
 
 # This override is due to this delightful cluster:
